StorageNode and Domain become ISessionSource-es
