+++ Orm/Xtensive.Orm/Orm/Upgrade/UpgradingDomainBuilder.cs
@@ -328,12 +328,13 @@ namespace Xtensive.Orm.Upgrade
       var modelMapping = ModelMappingBuilder.Build(
         domain.Handlers, schemaExtractionResult,
         context.Services.MappingResolver, context.NodeConfiguration, context.UpgradeMode.IsLegacy());
-      var result = new StorageNode(context.NodeConfiguration, modelMapping, new TypeIdRegistry());
+      var result = new StorageNode(domain, context.NodeConfiguration, modelMapping, new TypeIdRegistry());
 
       // Register default storage node immediately,
       // non-default nodes are registered in NodeManager after everything completes successfully.
-      if (result.Id==WellKnown.DefaultNodeId)
-        domain.Handlers.StorageNodeRegistry.Add(result);
+      if (result.Id==WellKnown.DefaultNodeId) {
+        _ = domain.Handlers.StorageNodeRegistry.Add(result);
+      }
 
       context.StorageNode = result;
       return result;
